backend/requirements.txt
fastapi==0.110.0
uvicorn==0.28.0
pydantic==2.6.1
motor==3.3.2
python-jose==3.3.0
passlib==1.7.4
python-dotenv==1.0.1
python-multipart==0.0.9
bcrypt==4.1.2